simplify: flatten doubled picker dir, drop dupe factory + dead code
- Flatten matches/opponent-picker/opponent-picker/ → the wrapper now sits
  as opponent-picker.tsx beside its opponent-picker/ deps dir, matching the
  scoreboard/match-list convention.
- Collapse buildOptionPlayer into the shared buildPlayer factory.
- Remove unused page-object methods (queryListbox/queryAlert) and dead
  type re-exports (Player, PlayerRead).
- Comment the typeahead's open-on-mount state.
