+++ web-client/src/mocks/factories/players/player.factory.ts
@@ -14,5 +14,3 @@ export function buildPlayer(overrides: Partial<PlayerRead> = {}): PlayerRead {
     ...overrides,
   }
 }
-
-export type { PlayerRead }
